Se modifico base de datos para le empresa y sus datos y tambien se ha avanzado en la generacion del JSON al pagar y generara factura.

+++ FacturacionElectronica/CLS/cuerpoDocumento.cs
@@ -16,14 +16,14 @@ namespace FacturacionElectronica.CLS
         public string descripcion { get; set; }
         public int cantidad { get; set; }
         public int uniMedida { get; set; }
-        public int precioUni { get; set; }
-        public int montoDescu { get; set; }
+        public Double precioUni { get; set; }
+        public Double montoDescu { get; set; }
         public int ventaNoSuj { get; set; }
         public int ventaExenta { get; set; }
-        public int ventaGravada { get; set; }
+        public Double ventaGravada { get; set; }
         public object tributos { get; set; }
         public double psv { get; set; }
         public int noGravado { get; set; }
-        public int ivaItem { get; set; }
+        public Double ivaItem { get; set; }
     }
 }
