+++ FacturacionElectronica/CLS/pagos.cs
@@ -9,7 +9,7 @@ namespace FacturacionElectronica.CLS
     public class pagos
     {
         public string codigo { get; set; }
-        public int montoPago { get; set; }
+        public Double montoPago { get; set; }
         public object plazo { get; set; }
         public string referencia { get; set; }
         public object periodo { get; set; }
